+++ src/BufferCopy/RDRAMtoColorBuffer.cpp
@@ -280,7 +280,7 @@ void RDRAMtoColorBuffer::copyFromRDRAM(u32 _address, bool _bCFB)
 
 	CombinerInfo::get().updateParameters();
 
-	glBindFramebuffer(GL_DRAW_FRAMEBUFFER, m_pCurBuffer->m_FBO);
+	glBindFramebuffer(GL_DRAW_FRAMEBUFFER, GLuint(m_pCurBuffer->m_FBO));
 
 	GraphicsDrawer::TexturedRectParams params((float)x0, (float)y0, (float)width, (float)height,
 										 0.0f, 0.0f, width - 1.0f, height - 1.0f, 1.0f, 1.0f,
