+++ src/windows/OpenGL_windows.cpp
@@ -131,9 +131,9 @@ void OGLVideoWindows::_saveScreenshot()
 	else {
 		if (config.video.multisampling != 0) {
 			pBuffer->resolveMultisampledTexture();
-			glBindFramebuffer(GL_READ_FRAMEBUFFER, pBuffer->m_resolveFBO);
+			glBindFramebuffer(GL_READ_FRAMEBUFFER, GLuint(pBuffer->m_resolveFBO));
 		} else
-			glBindFramebuffer(GL_READ_FRAMEBUFFER, pBuffer->m_FBO);
+			glBindFramebuffer(GL_READ_FRAMEBUFFER, GLuint(pBuffer->m_FBO));
 		pixelData = (unsigned char*)malloc(pBuffer->m_pTexture->realWidth * pBuffer->m_pTexture->realHeight * 3);
 		glReadPixels(0, 0, pBuffer->m_pTexture->realWidth, pBuffer->m_pTexture->realHeight, GL_RGB, GL_UNSIGNED_BYTE, pixelData);
 		SaveScreenshot(m_strScreenDirectory, RSP.romname, pBuffer->m_pTexture->realWidth, pBuffer->m_pTexture->realHeight, pixelData);
